# docs/swagger.yaml
basePath: /api
definitions:
...
  global.MODULE:
    properties:
      active:
        type: string
      comment:
        type: string
      create_date:
        type: string
      create_user:
        type: integer
      icon:
        type: string
      irl:
        type: string
      module_id:
        type: integer
      module_name:
        type: string
      module_seq:
        type: string
      parent_module:
        type: string
      update_date:
        type: string
      update_user:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a Api Documentation for SIPIL 2022 with Golang Backend, Gin
    Gonic Framework, GORM MySQL, and JWT Authentication
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API DOCUMENTATION - API FOR SIPIL 2022
  version: "1.0"
paths:
...
  /taskDTL:
    get:
      consumes:
      - application/json
      description: API untuk mengambil semua taskDTLyang terdapat dalam database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Mengambil Semua TaskDTLs
      tags:
      - TaskDTLs
    post:
      consumes:
      - application/json
      description: API untuk membuat taskDTLbaru.
      parameters:
      - description: TaskDTLs Data
        in: body
        name: taskDTL
        required: true
        schema:
          $ref: '#/definitions/global.MODULE'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Membuat taskDTLbaru.
      tags:
      - TaskDTLs
  /taskDTL/{id}:
    delete:
      consumes:
      - application/json
      description: API untuk delete taskDTL by ID
      parameters:
      - description: TaskDTL ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Delete taskDTL
      tags:
      - TaskDTLs
    get:
      consumes:
      - application/json
      description: API untuk mencari taskDTLby ID yang terdapat dalam database.
      parameters:
      - description: TaskDTLs ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Mengambil TaskDTLs by ID
      tags:
      - TaskDTLs
    put:
      consumes:
      - application/json
      description: API untuk update taskDTL
      parameters:
      - description: TaskDTLs Data
        in: body
        name: taskDTL
        required: true
        schema:
          $ref: '#/definitions/global.MODULE'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - bearerAuth: []
      summary: Update taskDTL
      tags:
      - TaskDTLs
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"